refactored scripts for build in docker containers

+++ linux/debian/scripts/compile_and_pack.sh
@@ -1,26 +1,8 @@
 #!/bin/bash
 
-#if [ "$1" == "--static" ]; then
-#	export $QT_SELECT="qtstatic" #For static builds we'll have a special qt instance, which should be installed manually for now, unfortunately.
-#fi
-
-# error_explainer() {
-
-#	case "$1" in
-#		"1"	) echo "Error in pre-config happened. Please, review logs";;
-#		"2"	) echo "Error in compilation happened. Please, review logs";;
-#		*	) echo "Unandled error $1 happened. Please, review logs";;
-#	esac
-#	exit $1
-
-# }
 
 
 substitute_pkgname_postfix() {
-
-	#CODENAME=$(lsb_release -a | grep Codename | cut -f2)
-	#VERSION=$(lsb_release -a | grep Version | cut -f2)
-	#DISTRO_TYPE=$(lsb_release -a | grep Distributor | cut -f2)
 	for variable in $(lsb_release -a 2>/dev/null | sed 's/\t//g' | sed 's/ //g' | sed 's/\:/\=/g'); do
 		echo "variable is $variable"
 		export $variable
@@ -28,7 +10,6 @@ substitute_pkgname_postfix() {
 	sed -i "/ CPACK_SYSTEM_TYPE/s/\".*\"/\"$DistributorID\"/" CMakeLists.txt
 	sed -i "/ CPACK_SYSTEM_VERSION/s/\".*\"/\"$Release\"/" CMakeLists.txt
 	sed -i "/ CPACK_SYSTEM_CODENAME/s/\".*\"/\"$Codename\"/" CMakeLists.txt
-#	sed -i "/ CPACK_SYSTEM_ARCH/s/\".*\"/\"$Codename\"/" CMakeLists.txt No need to replace anything while we're on amd64 arch only.
 	export -n "DistributorID"
 	export -n "Release"
 	export -n "Codename"
@@ -65,10 +46,25 @@ mkdir -p packages
 substitute_pkgname_postfix && mkdir -p build && cd build && cmake ../ && make -j$(nproc) && cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * \
 && cmake -DCMAKE_BUILD_TYPE=Debug ../ && make -j$(nproc) && cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * || error=$?
 
-sed -ibak 's/#set(BUILD_WITH_GDB_DRIVER_PGSQL ON)/set(BUILD_WITH_GDB_DRIVER_PGSQL ON)/' ../CMakeLists.txt || error=$?
-cmake ../ && make -j$(nproc) && cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * \
-&& cmake -DCMAKE_BUILD_TYPE=Debug ../ && make -j$(nproc) && cpack && repack *.deb && mv -v *.deb ../packages/ && cd .. && rm -r build && \
-mv -f CMakeLists.txtbak CMakeLists.txt || error=$?
+if [[ $(echo $ARCH_VERSIONS | grep arm64) != "" && $error == 0 ]]; then
+	cmake -DCMAKE_C_COMPILER=$ARM64_C_COMPILER -DCMAKE_CXX_COMPLIER=$ARM64_CXX_COMPILER -DCMAKE_TARGET_ARCH="arm64" .. && make -j$(nproc) && \
+	cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * || error=$?
+done
+
+if [[ $(echo $ARCH_VERSIONS | grep armhf) != "" && $error == 0 ]]; then
+	cmake -DCMAKE_C_COMPILER=$ARMHF_C_COMPILER -DCMAKE_CXX_COMPLIER=$ARMHF_CXX_COMPILER -DCMAKE_TARGET_ARCH="armhf" .. && make -j$(nproc) && \
+	cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * || error=$?
+done
+
+if [[ $(echo $BUILD_TYPE | grep PGSQL) != "" && $error == 0 ]]; then
+	sed -ibak 's/#set(BUILD_WITH_GDB_DRIVER_PGSQL ON)/set(BUILD_WITH_GDB_DRIVER_PGSQL ON)/' ../CMakeLists.txt || error=$?
+	cmake ../ && make -j$(nproc) && cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * \
+	&& cmake -DCMAKE_BUILD_TYPE=Debug ../ && make -j$(nproc) && cpack && repack *.deb && mv -v *.deb ../packages/ && rm -r * || error=$?
+done
+
+cd .. && rm -r build
+[ -e CMakeLists.txtbak ] && mv -f CMakeLists.txtbak CMakeLists.txt
+
 exit $error
 
 
@@ -77,8 +73,6 @@ exit $error
 ### wget https://debian.pub.demlabs.net/debian.pub.demlabs.net.gpg
 ### apt-key add demlabskey.asc
 
-### apt-get update
-### apt-get install cellframe-node
